--- Shared/ApiModel/SubcategoryApiModel.cs
@@ -1,10 +1,0 @@
-﻿namespace ValhallaVaultCyberAwareness.DAL.ApiModels
-{
-    public class SubcategoryApiModel
-    {
-        public int Id { get; set; }
-        public string Name { get; set; } = null!;
-        public int SegmentId { get; set; }
-        public List<QuestionApiModel> Questions { get; set; } = new();
-    }
-}
