+++ ValhallaVaultCyberAwareness/Components/Pages/Category.razor
@@ -2,7 +2,6 @@
 @using Microsoft.AspNetCore.Authorization
 @using Microsoft.AspNetCore.Mvc
 @using ValhallaVaultCyberAwareness.App.Services
-@using ValhallaVaultCyberAwareness.DAL.ApiModel
 @using ValhallaVaultCyberAwareness.DAL.DbModels
 @using ValhallaVaultCyberAwareness.API
 @using ValhallaVaultCyberAwareness.DAL.Repository
